Isolate saved query types
